requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=10.0.0
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
google-generativeai>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
pytest>=7.4.0
streamlit>=1.30.0
